Implemented default value settings for Link
+ Now Link, when created, will automatically be set with createdAt and
  updatedAt values
+ Now Link will auto update updatedAt value when updated
+ New URL extention to get URL of CoreData database
