+++ Link+CoreDataClass.swift
@@ -1,0 +1,15 @@
+//
+//  Link+CoreDataClass.swift
+//  Ulry
+//
+//  Created by Mattia Righetti on 1/12/22.
+//
+//
+
+import Foundation
+import CoreData
+
+
+public class Link: NSManagedObject {
+
+}
