+++ Ulry/Shared/Persistence.swift
@@ -7,17 +7,6 @@
 
 import CoreData
 
-public extension URL {
-    /// Returns a URL for the given app group and database pointing to the sqlite database.
-    static func storeURL(for appGroup: String, databaseName: String) -> URL {
-        guard let fileContainer = FileManager.default.containerURL(forSecurityApplicationGroupIdentifier: appGroup) else {
-            fatalError("Shared file container could not be created.")
-        }
-
-        return fileContainer.appendingPathComponent("\(databaseName).sqlite")
-    }
-}
-
 struct PersistenceController {
     static let shared = PersistenceController()
